finish post component
